+++ projects/gov-ui/src/lib/components/gov-uk-error-summary/gov-uk-error-summary.component.ts
@@ -16,7 +16,9 @@ import {DOCUMENT} from '@angular/common';
               <div class="govuk-error-summary__body">
                 <ul class="govuk-list govuk-error-summary__list">
                   <li *ngFor="let message of messages">
-                    <a (click)="scrollTo(message['id'])" [routerLink]="" [fragment]="message['id']">{{message['message']}}</a>
+                    <a (click)="scrollTo(message['id'])" [routerLink]="" [fragment]="message['id']"
+                      *ngIf="hasElement(message['id'])">{{message['message']}}</a>
+                    <span *ngIf="!hasElement(message['id'])">{{message['message']}}</span>
                   </li>
                 </ul>
               </div>
@@ -27,11 +29,11 @@ export class GovUkErrorSummaryComponent implements AfterViewInit, OnChanges {
 
   @Input() set errorMessages(value) {
     this.messages = value;
-  };
+  }
 
   messages: object[];
 
-  constructor (@Inject(DOCUMENT) private document) { }
+  constructor(@Inject(DOCUMENT) private document) { }
 
   ngAfterViewInit(): void {
     this.scrollTo('errorSummary');
@@ -49,6 +51,10 @@ export class GovUkErrorSummaryComponent implements AfterViewInit, OnChanges {
 
       el.focus();
     }
+  }
+
+  hasElement(selector) {
+    return this.document.querySelector(`#${selector}`);
   }
 
 }
