+++ src/users/store/actions/user.actions.ts
@@ -1,6 +1,5 @@
 // load login form
 import { Action } from '@ngrx/store';
-import {UserListApiModel} from '../../models/userform.model';
 
 export const LOAD_USERS = '[User] Load Users';
 export const LOAD_USERS_SUCCESS = '[User] Load Users Success';
@@ -22,32 +21,7 @@ export class LoadUsersFail implements Action {
   constructor(public payload: any) { }
 }
 
-export const SEND_INVITE_USER = '[User] Invite Users';
-export const INVITE_USER_SUCCESS = '[User] Invite Users Success';
-export const INVITE_USER_FAIL = '[User] Invite Users Fail';
-
-
-
-export class SendInviteUser {
-  readonly type = SEND_INVITE_USER;
-  constructor(public payload: UserListApiModel) {}
-}
-
-export class InviteUserSuccess implements Action {
-  readonly type = INVITE_USER_SUCCESS;
-  constructor(public payload: any) { }  // TODO add type list of users
-}
-
-export class InviteUserFail implements Action {
-  readonly type = INVITE_USER_FAIL;
-  constructor(public payload: any) { }
-}
-
-
 export type UserActions =
   | LoadUsers
   | LoadUsersSuccess
-  | LoadUsersFail
-  | SendInviteUser
-  | InviteUserSuccess
-  | InviteUserFail;
+  | LoadUsersFail;
