+++ src/users/store/reducers/invite-user.reducer.ts
@@ -50,6 +50,21 @@ export function reducer(
       };
     }
 
+    case fromInviteUsers.INVITE_USER_FAIL: {
+      const formErrorMessages = {
+        serverResponse: {
+          messages: [
+            'Try again later.'
+          ]
+        }
+      };
+      return {
+        ...state,
+        formErrorMessages,
+        isFormValid: false
+      };
+    }
+
   }
 
   return state;
